Changed the field by witch the photos are ordered in OrderedGallery.in_order() method
Due to the fact that PhotoOrdering.order field value never increases and
all entries have the same value (9999) the ordering in templates does not
work.
This is why I used PhotoOrdering's PK field to order the Photos. PK always
increases and is consistent with the order the PhotoOrdering objects were
defined in the Admin. This will have to do until proper (re)ordering is
implemented.